secure HTTP Server

+++ NodeJS/conFusionServer/routes/dishRouter.js
@@ -14,7 +14,9 @@ dishRouter.use(bodyParser.json());
 dishRouter.route('/')
 .get((req,res,next) => {
     Dishes.find({})
+    .populate('comments.author')
     .then((dishes) => {
+        //authenticate.verifyAdmin
         res.statusCode = 200;
         res.setHeader('Content-Type', 'application/json');
         res.json(dishes);
@@ -22,6 +24,7 @@ dishRouter.route('/')
     .catch((err) => next(err));
 })
 .post(authenticate.verifyUser, (req, res, next) => {
+    authenticate.verifyAdmin(req.user.admin)
     Dishes.create(req.body)
     .then((dish) => {
         console.log('Dish Created ', dish);
@@ -32,10 +35,12 @@ dishRouter.route('/')
     .catch((err) => next(err));
 })
 .put(authenticate.verifyUser, (req, res, next) => {
+    authenticate.verifyAdmin(req.user.admin)
     res.statusCode = 403;
     res.end('PUT operation not supported on /dishes');
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
+    authenticate.verifyAdmin(req.user.admin)
     Dishes.remove({})
     .then((resp) => {
         res.statusCode = 200;
@@ -48,6 +53,7 @@ dishRouter.route('/')
 dishRouter.route('/:dishId')
 .get((req,res,next) => {
     Dishes.findById(req.params.dishId)
+    .populate('comments.author')
     .then((dish) => {
         res.statusCode = 200;
         res.setHeader('Content-Type', 'application/json');
@@ -55,11 +61,14 @@ dishRouter.route('/:dishId')
     }, (err) => next(err))
     .catch((err) => next(err));
 })
+
 .post(authenticate.verifyUser, (req, res, next) => {
+    authenticate.verifyAdmin(req.user.admin)
     res.statusCode = 403;
     res.end('POST operation not supported on /dishes/'+ req.params.dishId);
 })
 .put(authenticate.verifyUser, (req, res, next) => {
+    authenticate.verifyAdmin(req.user.admin)
     Dishes.findByIdAndUpdate(req.params.dishId, {
         $set: req.body
     }, { new: true })
@@ -71,6 +80,7 @@ dishRouter.route('/:dishId')
     .catch((err) => next(err));
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
+    authenticate.verifyAdmin(req.user.admin)
     Dishes.findByIdAndRemove(req.params.dishId)
     .then((resp) => {
         res.statusCode = 200;
@@ -82,7 +92,9 @@ dishRouter.route('/:dishId')
 
 dishRouter.route('/:dishId/comments')
 .get((req,res,next) => {
+    authenticate.verifyAdmin(req.user.admin)
     Dishes.findById(req.params.dishId)
+    .populate('comments.author')
     .then((dish) => {
         if (dish != null) {
             res.statusCode = 200;
@@ -101,12 +113,17 @@ dishRouter.route('/:dishId/comments')
     Dishes.findById(req.params.dishId)
     .then((dish) => {
         if (dish != null) {
+            req.body.author = req.user._id;
             dish.comments.push(req.body);
             dish.save()
             .then((dish) => {
-                res.statusCode = 200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(dish);                
+                Dishes.findById(dish._id)
+                .populate('comments.author')
+                .then((dish) => {
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type', 'application/json');
+                    res.json(dish);
+                })            
             }, (err) => next(err));
         }
         else {
@@ -148,6 +165,7 @@ dishRouter.route('/:dishId/comments')
 dishRouter.route('/:dishId/comments/:commentId')
 .get((req,res,next) => {
     Dishes.findById(req.params.dishId)
+    .populate('comments.author')    
     .then((dish) => {
         if (dish != null && dish.comments.id(req.params.commentId) != null) {
             res.statusCode = 200;
@@ -173,6 +191,12 @@ dishRouter.route('/:dishId/comments/:commentId')
         + '/comments/' + req.params.commentId);
 })
 .put(authenticate.verifyUser, (req, res, next) => {
+    console.log(req.user.id)
+    if (req.params.commentId.equals(req.user.id)){
+        err = new Error('You are not authorized to perform this operation');
+        err.status = 403;
+        throw err;
+    }
     Dishes.findById(req.params.dishId)
     .then((dish) => {
         if (dish != null && dish.comments.id(req.params.commentId) != null) {
@@ -184,9 +208,13 @@ dishRouter.route('/:dishId/comments/:commentId')
             }
             dish.save()
             .then((dish) => {
-                res.statusCode = 200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(dish);                
+                Dishes.findById(dish._id)
+                .populate('comments.author')
+                .then((dish) => {
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type', 'application/json');
+                    res.json(dish);  
+                })              
             }, (err) => next(err));
         }
         else if (dish == null) {
@@ -203,15 +231,26 @@ dishRouter.route('/:dishId/comments/:commentId')
     .catch((err) => next(err));
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
+    commentIdd = req.params.commentId
+    userIdd = req.user.id
+    if (commentIdd != userIdd){
+        err = new Error('You are not authorized to perform this operation');
+        err.status = 403;
+        throw err;
+    }
     Dishes.findById(req.params.dishId)
     .then((dish) => {
         if (dish != null && dish.comments.id(req.params.commentId) != null) {
             dish.comments.id(req.params.commentId).remove();
             dish.save()
             .then((dish) => {
-                res.statusCode = 200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(dish);                
+                Dishes.findById(dish._id)
+                .populate('comments.author')
+                .then((dish) => {
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type', 'application/json');
+                    res.json(dish);  
+                })               
             }, (err) => next(err));
         }
         else if (dish == null) {
